Replace the half-baked build system I was winging with one adapted from the FFMS2 C plugin. May contain some wonky or broken functionality for odd cases, but the 'meat' of it should work, and it can always be tailored better at a later point.
Also fix the previous ASM merge into GNUmakefile. It didn't merge in
any of the stuff pointing to where or what the ASM_OBJECTS were, which
was also where the AR and RANLIB parts were.  The result was that the
CXX objects would compile, but it wouldn't compile the ASM or link
anything anymore.  It should work correctly now.