Fix shared compilation on MinGW by ifdef-ing around Format.o
It's not a proper solution, especially since this removes it from
the static library too (and the static library didn't have a problem
with it).

Also fixes the install routines for shared compilation across the
board, and tidies up clean, uninstall, and distclean some. The only
thing left is the weird treatment the static library gets during
'make install' if --enable-shared was used.